0.4.0 work
* Added docs suproject with tut and such for GH pages
* Added generic subproject with shapeless support and a minimal query DSL
* A lot of refactoring of main project
